// MultiSelectFlatCB/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    fieldValue: ComponentFramework.PropertyTypes.StringProperty;
    entityNameToTransform: ComponentFramework.PropertyTypes.StringProperty;
    fieldName: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    fieldValue?: string;
}